Scrolling works, but WHAT?! DAFUQ
I can now use ↑ ↓ → ← keys to scroll the background.  Of course if I go ← the background goes the other way.
If you scroll off the background.. WHAT?  the background looks duplicated.
This is because I did not clear the screen on every loop, so when I scroll the background to far, part of the window is not touched by the drawing of the background picture.
So you are seeing pixels that were drawn last loop but now is untouched.  I better clear the screen every loop!